feat: meeting page layout 구조 변경 및 video list 분리
