+++ app/(root)/workspace/meeting/[meetingId]/layout.tsx
@@ -1,7 +1,5 @@
 import { ReactNode } from 'react';
 
-import ButtonGroup from '@/components/meeting/meetingRoom/ButtonGroup';
-import MeetingTitle from '@/components/meeting/meetingRoom/MeetingTitle';
 import { MeetingProvider } from '@/lib/websocket/MeetingProvider';
 
 const MeetingRoomLayout = ({
@@ -11,11 +9,7 @@ const MeetingRoomLayout = ({
 }>) => {
 	return (
 		<div className="flex flex-col justify-between h-full">
-			<MeetingTitle />
-			<MeetingProvider>
-				{children}
-				<ButtonGroup />
-			</MeetingProvider>
+			<MeetingProvider>{children}</MeetingProvider>
 		</div>
 	);
 };
